Seperate `courses/` endpoints out of main.py
